Fix missing constant when running artifactory_sha1 spec alone
Under parallel_tests the artifactory_sha1_spec may run alone, but for it
to set up a stub on PuppetX::Bodeco::Util the util file must be loaded
first. Because the 'require' was in the body of the function, it would
be loaded too late, only when the function was run.