Produce a better error for the puppet downloader when file not found
Newer versions of Puppet raise the error Puppet::HTTP::RouteError when the file is not found. This resulted not only in a misleading error but also an error without the original url.

This commit changes this. It catches the error and produces an error with the url  in it. Unfortunately, the original misleading error is still produces, so we try to elaborate a bit in the error we produce.